+++ src/briefy/leica/views/dashboard/__init__.py
@@ -10,10 +10,6 @@ ORDER_PROJECT_COLS = (
         'type': 'text', 'url': 'absolute_url', 'filter': ''
     },
     {
-        'field': 'total', 'label': 'All Orders',
-        'type': 'integer', 'url': '', 'filter': ''
-    },
-    {
         'field': 'received', 'label': 'New Orders',
         'type': 'integer', 'url': '', 'filter': ''
     },
@@ -26,7 +22,7 @@ ORDER_PROJECT_COLS = (
         'type': 'integer', 'url': '', 'filter': ''
     },
     {
-        'field': 'in_qa', 'label': 'Orders in QA process',
+        'field': 'in_qa', 'label': 'In Briefy QA',
         'type': 'integer', 'url': '', 'filter': ''
     },
     {
@@ -34,7 +30,42 @@ ORDER_PROJECT_COLS = (
         'type': 'integer', 'url': '', 'filter': ''
     },
     {
-        'field': 'completed', 'label': 'Completed Orders',
+        'field': 'delivered', 'label': 'Delivered Orders',
         'type': 'integer', 'url': '', 'filter': ''
-    }
+    },
+    {
+        'field': 'total', 'label': 'Total',
+        'type': 'integer', 'url': '', 'filter': ''
+    },
+)
+
+DELIVERED_ORDERS_COLS = (
+    {
+        'field': 'absolute_url', 'label': 'url',
+        'type': 'hidden', 'url': '', 'filter': ''
+    },
+    {
+        'field': 'title', 'label': 'Project',
+        'type': 'text', 'url': 'absolute_url', 'filter': ''
+    },
+    {
+        'field': 'newly-delivered', 'label': 'Newly Delivered',
+        'type': 'integer', 'url': '', 'filter': ''
+    },
+    {
+        'field': 'further_revision', 'label': 'In Further Revision',
+        'type': 'integer', 'url': '', 'filter': ''
+    },
+    {
+        'field': 're-delivered', 'label': 'Re-delivered',
+        'type': 'integer', 'url': '', 'filter': ''
+    },
+    {
+        'field': 'completed', 'label': 'Completed',
+        'type': 'integer', 'url': '', 'filter': ''
+    },
+    {
+        'field': 'total', 'label': 'Total',
+        'type': 'integer', 'url': '', 'filter': ''
+    },
 )