+++ HISTORY.rst
@@ -2,6 +2,17 @@
 History
 =======
 
+2.1.1 (2017-03-29)
+------------------
+
+    * Fix: remove_availability transition now create a new assignment before cancel the old one (rudaporto).
+    * Fix: when QA approve a set, creative comment was not being created as a comment in the Assignment (rudaporto).
+    * Card #132: Added new _custom_filter to Orders endpoint to be used by the 'Deliveries' tab in customer interface (rudaporto).
+    * Card #128: New dashboard for Customer and PM: delivered (rudaporto).
+    * Fix: perm_rejected transitions to edit payout and edit compensation typo in definition (rudaporto).
+    * Card #155: Update All Orders dashboard for PM, Customer and Bizdev (rudaporto).
+    * Card #157: Update Orders export csv with new label for each workflow state (rudaporto).
+
 2.1.0 (2017-03-26)
 ------------------